% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser_bind}
\alias{parser_bind}
\alias{parser_then}
\alias{parser_return}
\alias{parser_fail}
\alias{parser_do}
\title{Parser Sequencing}
\usage{
parser_bind(p, f)

parser_then(p, q)

parser_return(value)

parser_fail(error)

parser_do(...)
}
\arguments{
\item{p, q}{Parser objects.}

\item{f}{A function that takes the value parsed by \code{p} and returns a new parser.}

\item{value}{The parse value that the new parser resolves to.}

\item{error}{The parse error that the new parser resolves to.}

\item{...}{Parser objects.}
}
\value{
A new parser object.
}
\description{
\code{parser_return()} creates a parser that will resolve to the specified \code{value}
without consuming any input.
}
