% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser_many}
\alias{parser_many}
\alias{parser_all}
\alias{parser_any}
\alias{\%|\%.parser}
\title{Parser Combination}
\usage{
parser_many(p)

parser_all(...)

parser_any(...)

\method{\%|\%}{parser}(p, q)
}
\arguments{
\item{p, q, ...}{Parser objects.}
}
\value{
A parser object.
}
\description{
Parser Combination
}
