% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser_map}
\alias{parser_map}
\alias{parser_call}
\title{Parser Transformation}
\usage{
parser_map(p, f)

parser_call(p, f)
}
\arguments{
\item{p}{A parser object.}

\item{f}{A function that takes the value(s) parsed by \code{p} and returns a new value.}
}
\value{
A new parser object.
}
\description{
Parser Transformation
}
\examples{
parse_letters() |> parser_map(toupper) |> parser_run("hello world")
}
