% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parser_run}
\alias{parser_run}
\title{Run a Parser}
\usage{
parser_run(p, x)
}
\arguments{
\item{p}{A parser object.}

\item{x}{Input to be parsed.}
}
\value{
A parser state object wrapped in a result object.
}
\description{
Run a Parser
}
