% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsers.R
\name{parsers}
\alias{parsers}
\alias{parse_string}
\alias{parse_regexp}
\alias{parse_digits}
\alias{parse_letters}
\alias{parse_whitespace}
\title{Basic Parsers}
\usage{
parse_string(string)

parse_regexp(pattern)

parse_digits()

parse_letters()

parse_whitespace()
}
\arguments{
\item{string}{A sequence of characters to match.}

\item{pattern}{A \link[=regexp]{regular expression} to match.}
}
\value{
A parser object.
}
\description{
Basic Parsers
}
